For #10426: Adds identifier to Glean for 24 hours
